chore: Move usePopup to useBrowser
