+++ source/hooks/index.ts
@@ -1,5 +1,4 @@
 export * from './useBrowser';
-export * from './usePopup';
 export * from './usePrice';
 export * from './useStore';
 export * from './useTransaction';
